backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
python-multipart==0.0.20
pydantic==2.11.7
psycopg[binary]==3.2.10
faster-whisper>=1.0.0,<2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-runtime-cu12==12.9.79
pytest==8.4.1
httpx==0.28.1
python-dotenv==1.0.1
numpy>=1.26,<3
sentence-transformers>=3.0.0,<4
spacy>=3.7,<4
